refactor: remove providedIn:root from  SandboxFetcherService
